Move DDSI configuration units to confgen tool
The configuration generation tool is the only place where these definitions are used,
moving them makes them easier to find.
