+++ .pre-commit-config.yaml
@@ -23,10 +23,10 @@ repos:
       '--hash-files',
         'src/core/ddsi/include/dds/ddsi/ddsi_config.h',
         'src/core/ddsi/src/ddsi__cfgelems.h',
-        'src/core/ddsi/src/ddsi__cfgunits.h',
         'src/core/ddsi/src/ddsi_config.c',
         'src/tools/_confgen/_confgen.h',
         'src/tools/_confgen/_confgen.c',
+        'src/tools/_confgen/cfgunits.h',
         'src/tools/_confgen/generate_rnc.c',
         'src/tools/_confgen/generate_md.c',
         'src/tools/_confgen/generate_xsd.c',
