+++ src/core/ddsi/CMakeLists.txt
@@ -144,7 +144,6 @@ set(hdrs_ddsi
 
 set(hdrs_private_ddsi
   ddsi__acknack.h
-  ddsi__cfgunits.h
   ddsi__cfgelems.h
   ddsi__config_impl.h
   ddsi__deliver_locally.h
